Final tally: 40 failed auto-generated tests out of 68
In conclusion, the factory is a really exciting idea, but its success
rate at creating tests for us automatically is less than 50%. As such,
I don’t think it will be a very useful tool. I’ll leave it in for now,
but I probably won’t be using it.