Made minor changes to the codebase to deal with a bunch of warnings
Suppressed a few warnings I couldn’t do anything about because they’re
part of Python 2.6.9, and changed our code in other places to appease
the warnings. We were using deprecated references in a few places, so I
followed up on the warning’s recommendations for the code updates. Also
I changed it so that Sentry will not log exceptions thrown in unit
testing.