In-Progress Date Delete
Currently will move the start and end date of the one deleting if it is
the first one being deleted, since I hard-coded the for loop to do that
for testing if data transfer worked.

So change for loop to start at ID starting at.

Still needs the unique ID of each delete ‘X’ to be established when
clicked on specific delete date in the javascript.

Also, need to transfer data from check boxes.

Think that’s it.