Added a few 'missing' unit test files and fixed some typos
The files hadn’t been added before this commit because I figured they
wouldn’t be tested. Caleb pointed out that making stand-in files for
the methods we aren’t testing and just passing on the relevant testing
methods doesn’t add a ton of overhead time and gives a place to explain
why that method isn’t being tested.