some small fixes to logic
moved counter to bottom of loop so it actually counts everything, including when it doesnt check that redirect. ALSO, moved sleep into the check so we don't have to sleep everytime the loop does its thing. Finally, switched the end logic to only break if the end position is actually met